requirements.txt
# PyTorch CPU-only (latest stable versions)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0

# Code quality
black>=23.0.0
flake8>=6.0.0
